services/mcp-server/requirements.txt
# Write the server requirements file (same pins we planned)
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
GitPython==3.1.43
requests==2.32.3
boto3==1.35.36
python-dotenv==1.0.1